moved logger functions from daq_utils to logger module
