+++ src/pymodaq/examples/function_plotter.py
@@ -9,6 +9,7 @@ import numpy as np
 from qtpy import QtWidgets
 from qtpy.QtCore import Slot, QDate, QThread, QTimer
 
+from pymodaq.utils.logger import set_logger
 from pymodaq.utils.gui_utils.custom_app import CustomApp
 from pymodaq.utils.gui_utils.dock import DockArea, Dock
 from pymodaq.utils.gui_utils.file_io import select_file
@@ -22,7 +23,7 @@ from pymodaq.utils.plotting.data_viewers.viewer1D import Viewer1D
 from pymodaq.utils.h5modules import H5Browser, H5Saver
 
 config = Config()
-logger = utils.set_logger(utils.get_module_name(__file__))
+logger = set_logger(utils.logger.get_module_name(__file__))
 
 
 class FunctionPlotter(CustomApp):
