+++ src/pymodaq/utils/daq_utils.py
@@ -3,7 +3,8 @@ import sys
 from collections import OrderedDict
 from ctypes import CFUNCTYPE
 
-from pymodaq.utils.config import get_set_config_path, get_set_preset_path, Config
+from pymodaq.utils.logger import set_logger
+from pymodaq.utils.config import get_set_preset_path, Config
 from pymodaq.utils.messenger import deprecation_msg
 if 'win32' in sys.platform:
     from ctypes import WINFUNCTYPE
@@ -12,16 +13,13 @@ import importlib
 import inspect
 import json
 
-import logging
 import functools
 import re
 import time
-from logging.handlers import TimedRotatingFileHandler
 from packaging import version as version_mod
 from pathlib import Path
 import pkgutil
 import traceback
-import warnings
 import numbers
 
 import numpy as np
@@ -38,7 +36,7 @@ else:
 
 from pymodaq.utils.exceptions import DataSourceError
 
-
+logger = set_logger('daq_utils')
 plot_colors = [(255, 255, 255), (255, 0, 0), (0, 255, 0), (0, 0, 255), (14, 207, 189), (207, 14, 166), (207, 204, 14)]
 config = Config()
 
@@ -80,52 +78,7 @@ def timer(func):
     return wrapper_timer
 
 
-def set_logger(logger_name, add_handler=False, base_logger=False, add_to_console=False, log_level=None,
-               logger_base_name='pymodaq', local_dir=None):
-    """defines a logger of a given name and eventually add an handler to it
 
-    Parameters
-    ----------
-    logger_name: (str) the name of the logger (usually it is the module name as returned by get_module_name
-    add_handler (bool) if True adds a TimedRotatingFileHandler to the logger instance (should be True if logger set from
-                main app
-    base_logger: (bool) specify if this is the parent logger (usually where one defines the handler)
-
-    Returns
-    -------
-    logger: (logging.logger) logger instance
-    See Also
-    --------
-    get_module_name, logging.handlers.TimedRotatingFileHandler
-    """
-    if not base_logger:
-        logger_name = f'{logger_base_name}.{logger_name}'
-
-    logger = logging.getLogger(logger_name)
-    log_path = get_set_config_path('log', local_dir=local_dir)
-    if add_handler:
-        if log_level is None:
-            log_level = config('general', 'debug_level')
-        logger.setLevel(log_level)
-        handler = TimedRotatingFileHandler(log_path.joinpath(f'{logger_base_name}.log'), when='midnight')
-        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-        handler.setFormatter(formatter)
-        logger.addHandler(handler)
-
-        logging.captureWarnings(True)
-        warnings.filterwarnings('default', category=DeprecationWarning)
-        warnings_logger = logging.getLogger("py.warnings")
-        warnings_logger.addHandler(handler)
-
-    if add_to_console:
-        console_handler = logging.StreamHandler()
-        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-        console_handler.setFormatter(formatter)
-        logger.addHandler(console_handler)
-    return logger
-
-
-logger = set_logger('daq_utils')
 
 
 def get_version():
@@ -1163,12 +1116,6 @@ def check_vals_in_iterable(iterable1, iterable2):
     iterable2 = list(iterable2)
     for val1, val2 in zip(iterable1, iterable2):
         assert val1 == val2
-
-
-def get_module_name(module__file__path):
-    """from the full path of a module extract its name"""
-    path = Path(module__file__path)
-    return path.stem
 
 
 def caller_name(skip=2):
