+++ src/pymodaq/utils/managers/batchscan_manager.py
@@ -6,6 +6,8 @@ from qtpy import QtWidgets, QtCore
 import sys
 import os
 from pyqtgraph.parametertree import Parameter, ParameterTree
+
+from pymodaq.utils.logger import set_logger, get_module_name
 from pymodaq.utils.parameter import ioxml
 from pymodaq.utils.parameter import pymodaq_ptypes
 from pymodaq.utils.managers import preset_manager_utils
@@ -16,7 +18,7 @@ from pymodaq.utils.scanner import Scanner, SCAN_TYPES, adaptive_losses
 from pathlib import Path
 from collections import OrderedDict
 
-logger = utils.set_logger(utils.get_module_name(__file__))
+logger = set_logger(get_module_name(__file__))
 
 batch_path = pymodaq.utils.config.get_set_batch_path()
