+++ src/pymodaq/utils/plotting/scan_selector.py
@@ -6,11 +6,13 @@ from pyqtgraph import ROI, RectROI, PolyLineROI, Point
 
 from pyqtgraph.parametertree import Parameter, ParameterTree
 from pyqtgraph.dockarea import Dock
+
+from pymodaq.utils.logger import set_logger, get_module_name
 from pymodaq.utils.gui_utils import DockArea
 from pymodaq.utils.plotting.utils.plot_utils import QVector
 from pymodaq.utils import daq_utils as utils
 
-logger = utils.set_logger(utils.get_module_name(__file__))
+logger = set_logger(get_module_name(__file__))
 
 
 class PolyLineROI_custom(PolyLineROI):
